Move exception.h.
