+++ src/include/binder/binder_util.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "common/managed_pointer.h"
 #include "type/type_id.h"
