+++ src/include/catalog/postgres/name_builder.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "network/postgres/postgres_defs.h"
 
 namespace terrier::catalog::postgres {
