+++ src/include/network/connection_handler_task.h
@@ -5,11 +5,12 @@
 #include <event2/event.h>
 #include <event2/listener.h>
 #include <unistd.h>
+
 #include <deque>
 #include <memory>
 #include <utility>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "common/notifiable_task.h"
 #include "network/network_defs.h"
 #include "network/protocol_interpreter.h"
