+++ src/include/network/postgres/postgres_protocol_util.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <utility>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "network/network_defs.h"
 #include "network/postgres/postgres_defs.h"
