+++ src/include/network/terrier_server.h
@@ -8,14 +8,16 @@
 #include <netinet/tcp.h>
 #include <pthread.h>
 #include <sys/file.h>
+
 #include <csignal>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
 #include <memory>
 #include <vector>
+
 #include "common/dedicated_thread_owner.h"
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "common/notifiable_task.h"
 #include "network/connection_dispatcher_task.h"
 #include "network/connection_handle_factory.h"
