+++ src/include/optimizer/expression_node_contents.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "common/managed_pointer.h"
 #include "optimizer/abstract_optimizer_node_contents.h"
 #include "parser/expression/abstract_expression.h"
