+++ src/include/parser/expression/aggregate_expression.h
@@ -3,7 +3,8 @@
 #include <memory>
 #include <utility>
 #include <vector>
-#include "common/exception.h"
+
+#include "common/error/exception.h"
 #include "parser/expression/abstract_expression.h"
 #include "parser/expression_defs.h"
