+++ test/parser/expression_defs_test.cpp
@@ -1,7 +1,8 @@
 #include "parser/expression_defs.h"
+
 #include <string>
 
-#include "common/exception.h"
+#include "common/error/exception.h"
 #include "gtest/gtest.h"
 
 namespace terrier::parser {
